Additional Refactoring
- Renamed `Factory` namespace to `Factories`
- Refactored retrieval into a Repository
